+++ src/main.cpp
@@ -68,8 +68,9 @@ int main(int argc, char *argv[]) {
         po::options_description hidden("Hidden options");
 
         generic.add_options()
-            ("help,h", "produce help message")
-            ("version,v", "displays version of this program");
+            ("help,h",             "produce help message")
+            ("version,v",          "displays version of this program")
+            ("rebuild-cache,r",    "ignores cache file contents and instead queries SCIM server for list of objects");
 
         // Config file variables exposed as command line options
         std::vector<config_file_option> common_vars =
@@ -180,10 +181,20 @@ int main(int argc, char *argv[]) {
                 continue;
             }
 
+            SCIMServerInfo server_info{config};
+            ScimActions scim_actions{server_info};
 
             /** Get objects from cache file */
             std::shared_ptr<object_list> cache = cache_file::instance().get_contents();
-
+            std::vector<ScimActions::scim_object_ref> all_scim_objects;
+            if (vm.count("rebuild-cache")) {
+                try {
+                    all_scim_objects = scim_actions.get_all_objects_from_scim_server();
+                } catch (const std::runtime_error& e) {
+                    std::cerr << "Failed to get objects from SCIM server (" << e.what() << ")" << std::endl;
+                }
+            }
+            
             if (cache == nullptr) {
                 print_error();
                 server.clear();
@@ -193,7 +204,7 @@ int main(int argc, char *argv[]) {
 
             /** Perform SCIM operations */
             try {
-                err = ScimActions(SCIMServerInfo(config)).perform(server, *cache);
+                err = scim_actions.perform(server, *cache, vm.count("rebuild-cache"), all_scim_objects);
             } catch (const std::string& err_msg) {
                 std::cerr << err_msg << std::endl;
             }