+++ src/scim.hpp
@@ -49,6 +49,28 @@ struct variables {
 };
 
 class ScimActions {
+public:
+        /** A reference to an object in the SCIM server.
+     *  Contains a UUID and an endpoint.
+     *
+     *  This type is used when rebuilding the cache and fetch
+     *  all objects from the server. Since we fetch by endpoint,
+     *  and not all SS12000 types have different endpoints
+     *  (Students and Teachers both end up under Users), we
+     *  won't easily know what type of object it is. But need to
+     *  remember the endpoint so we can delete it if needed.
+     */
+    struct scim_object_ref {
+        std::string uuid;
+        std::string endpoint;
+
+        scim_object_ref(std::string u, std::string e)
+                : uuid(u),
+                  endpoint(e) {
+        }
+    };    
+
+private:
     std::shared_ptr<object_list> scim_new_cache;
     variables vars = variables();
     mutable string_vector verified_types;
@@ -67,13 +89,19 @@ class ScimActions {
     };    
     
     void process_changes(const object_list& current,
-                        const object_list& cache,
-                        statistics& stats) const;
+                         const object_list& cache,
+                         statistics& stats,
+                         bool rebuild_cache,
+                         const std::set<std::string>& all_scim_uuids) const;
 
     void process_deletes(const object_list& current,
                          const object_list& cache,
                          const std::string& type,
                          statistics& stats) const;
+
+    void process_deletes_per_endpoint(const std::vector<std::string>& to_delete,
+                                      const std::string& endpoint,
+                                      statistics& stats) const;    
 
     static void print_statistics(const std::string& type,
                                  const statistics& stats);
@@ -82,9 +110,39 @@ public:
     ScimActions(const SCIMServerInfo& si)
            : scim_server_info(si) {
         scim_new_cache = std::make_unique<object_list>();
+
+        int err = simplescim_scim_init();
+
+        if (err == -1) {
+            throw std::runtime_error("Failed to init SCIM");
+        }
     }
 
-    int perform(const data_server &current, const object_list &cached) const;
+    ~ScimActions() {
+        /* Clean up */
+        simplescim_scim_clear();
+    }    
+    
+    /**
+     * Makes SCIM requests by comparing the two user lists and
+     * reading JSON templates from the configuration file.
+     * Updates (or creates) the cache file.
+     *
+     * If rebuilding the cache (rebuild_cache = true), this
+     * function will ignore what's in cached and instead use
+     * all_scim_objects. For all objects in current, we will
+     * then either create or update (depending on if the object
+     * already exists in the SCIM server). All objects in
+     * all_scim_objects which aren't in current will be deleted.
+     *
+     * On success, zero is returned. On error, -1 is returned
+     * and simplescim_error_string is set to an appropriate
+     * error message.
+     */
+    int perform(const data_server &current,
+                const object_list &cached,
+                bool rebuild_cache,
+                const std::vector<scim_object_ref>& all_scim_objects) const;
     bool verify_json(const std::string &json, const std::string &type) const ;
 
     class copy_func {
@@ -105,10 +163,8 @@ public:
 
     class update_func {
         const base_object &object;
-//		const base_object &cached_object;
     public:
-        update_func(const base_object &o, const base_object &co) : object(o)
-//		, cached_object(co)
+        update_func(const base_object &o) : object(o)
             {}
 
         int operator()(const ScimActions &);
@@ -121,7 +177,14 @@ public:
 
         int operator()(const ScimActions &);
     };
-
+    
+    /** Gets a list of all resources in the SCIM server.
+     *  Will query the endpoints for the SS12000 types included in scim-type-send-order.
+     *
+     *  This function will throw std::runtime_error if we fail to get the resources
+     *  from the SCIM server.
+     */
+    std::vector<scim_object_ref> get_all_objects_from_scim_server();
 };